# api-gateway/api/docs/swagger.yaml
definitions:
...
  booking.ServiceCreateReq:
    properties:
      description:
        type: string
      duration:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  booking.ServiceGetAllRes:
    properties:
      services:
        items:
          $ref: '#/definitions/booking.ServiceRes'
        type: array
    type: object
  booking.ServiceGetByIdRes:
    properties:
      service:
        $ref: '#/definitions/booking.ServiceRes'
    type: object
  booking.ServiceRes:
    properties:
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /service:
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/booking.ServiceCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Service created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create service
      tags:
      - service
  /service/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Service deleted
          schema:
            type: string
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete service
      tags:
      - service
    get:
      consumes:
      - application/json
      description: Get a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get service
      tags:
      - service
    put:
      consumes:
      - application/json
      description: Update an existing service
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/booking.ServiceCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Service updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update service
      tags:
      - service
  /service/all:
    get:
      consumes:
      - application/json
      description: Get all service
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Price
        in: query
        name: price
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all service
      tags:
      - service
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"